Refactored helper methods out of use case
