+++ src/core/helpers/examData.helper.ts
@@ -8,7 +8,7 @@ import {
   ValidSensoryValues,
 } from '@core/domain';
 import {ExamData} from '@core/domain/examData';
-import {validCellNameRegex} from './regularExpressions';
+import {motorCellRegex, validCellNameRegex} from './regularExpressions';
 
 const validateValue = (
   dataKey: string,
@@ -203,4 +203,113 @@ export const findCell = (cellName: string, gridModel: Array<Cell | null>[]) => {
   }
 
   return cell;
+};
+
+/*
+ * Returns a number between 0 and 5 representing the column index of a cell in the grid model.
+ * Throws and error when the cell name is invalid.
+ */
+export const getCellColumn = (cellName: string) => {
+  if (!validCellNameRegex.test(cellName)) {
+    throw new Error(`Invalid cell name ${cellName}`);
+  }
+
+  switch (true) {
+    case /right-motor/.test(cellName):
+      return 0;
+    case /right-light-touch/.test(cellName):
+      return 1;
+    case /right-pin-prick/.test(cellName):
+      return 2;
+    case /left-light-touch/.test(cellName):
+      return 3;
+    case /left-pin-prick/.test(cellName):
+      return 4;
+    case /left-motor/.test(cellName):
+      return 5;
+    default:
+      throw new Error(`Invalid cell name ${cellName}`);
+  }
+};
+
+/*
+ * Returns a number between 0 and 27 representing the row index of a cell in the grid model.
+ * Throws and error if the cell name is invalid.
+ */
+export const getCellRow = (cellName: string) => {
+  const level = /((?!-)(c|t|l|s)1?[0-9]$)|((?!-)s4_5)/.exec(cellName)?.[0];
+
+  if (!level) {
+    throw new Error(`Invalid cell name ${cellName}`);
+  }
+
+  if (level === 's4_5') {
+    return 27;
+  }
+
+  const numericSegment = parseInt(level.slice(1));
+
+  switch (level[0]) {
+    case 'c':
+      return numericSegment - 2;
+    case 't':
+      return numericSegment + 6;
+    case 'l':
+      return numericSegment + 18;
+    case 's':
+      return numericSegment + 23;
+    default:
+      throw new Error(`Invalid cell name ${cellName}`);
+  }
+};
+
+/*
+ * Returns the column and row index of a cell in the grid model.
+ */
+export const getCellPosition = (cellName: string) => {
+  return {column: getCellColumn(cellName), row: getCellRow(cellName)};
+};
+
+/*
+ * Returns an array of cells between two cells in the grid model.
+ * The range is inclusive of the start and end cells.
+ */
+export const getCellRange = (
+  start: {column: number; row: number},
+  end: {column: number; row: number} | null,
+  gridModel: Array<Cell | null>[],
+  stopAtCellWithValue: boolean = false,
+) => {
+  const motorRange: Cell[] = [];
+  const sensoryRange: Cell[] = [];
+  const endPosition = end ?? {column: start.column, row: 27};
+
+  const rowStart = Math.min(start.row, endPosition.row);
+  const rowEnd = Math.max(start.row, endPosition.row);
+  const columnStart = Math.min(start.column, endPosition.column);
+  const columnEnd = Math.max(start.column, endPosition.column);
+
+  for (let row = rowStart; row <= rowEnd; row++) {
+    for (let column = columnStart; column <= columnEnd; column++) {
+      const cell = gridModel[row][column];
+      if (cell) {
+        if (
+          stopAtCellWithValue &&
+          end === null &&
+          cell.value &&
+          (row !== rowStart || column !== columnStart)
+        ) {
+          return {motorRange, sensoryRange};
+        }
+
+        if (motorCellRegex.test(cell.name)) {
+          motorRange.push(cell);
+        } else {
+          sensoryRange.push(cell);
+        }
+      }
+    }
+  }
+
+  return {motorRange, sensoryRange};
 };
